Stable + Missing Core

+++ Car Rental Service/Payment Service/src/main/java/Repository/IPaymentRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface IPaymentRepository {
   Boolean createPayment(Payment payment);
   List<Payment> getAllPayments();
-  Optional<Payment> getPaymentById(int id); //Optional indikere at værdien kan være null
+  Optional<Payment> getPaymentById(long id); //Optional indikere at værdien kan være null
   void updatePayment(Payment payment);
-  void deletePayment(int id);
+  void deletePayment(long id);
 }
